refactor(driver): move GRO/GSO consts + gro_recv_buf_len into transports/mod.rs
Addresses review feedback on #820 (rainliu): the UDP GRO/GSO batching
constants and the gro_recv_buf_len() helper describe the transport/socket
layer, not the driver's event loop, so they belong in
peer_connection::transports rather than in driver.rs.

Moved to transports/mod.rs (all pub(crate)):
  UDP_RECV_BUF_LEN, MAX_GRO_SEGMENTS, MAX_GSO_SEGMENTS, MAX_GSO_BATCH_BYTES,
  MIN_GSO_RUN, GRO_RECV_SEGMENT_LEN, fn gro_recv_buf_len, and its unit test.
driver.rs now imports them.

UDP_RECV_BUF_LEN moves too: it is the base UDP recv-buffer size and
gro_recv_buf_len falls back to it, so leaving it in driver would force a
backwards transports -> driver reference. transports is the lower layer here
(driver imports transports, never the reverse), so the const lives there.

No behavior change: the constant values and the fn body are byte-identical to
before the move (pure relocation), so there is no performance delta to measure
-- the GSO/GRO poop A/B tables stand unchanged in the parent commits (GRO recv
44f94250, GSO send d89c4629). Verified: tokio + smol build clean, clippy clean
on both runtimes, gro_recv_buf_len test passes in its new
peer_connection::transports::gro_buf_tests module.
